Refatoracao da estrutura da app backend para ficar modularizada
